+++ packages/client/src/helpers/getAvatarSrc.ts
@@ -1,0 +1,9 @@
+import { RESOURCES_URL } from '../api/basic/basicInstance'
+
+const DEFAULT_AVATAR = 'images/defaultAvatar.png'
+
+const getAvatarSrc = (avatar?: string) => {
+  return avatar ? `${RESOURCES_URL}${avatar}` : DEFAULT_AVATAR
+}
+
+export default getAvatarSrc
